requirements.txt
boto3==1.18.31
awsiotsdk==1.7.0
awsiot==0.1.3
moto==2.2.6
pylint==2.10.2
pytest==6.2.4
pytest-cov==2.12.1
pytest-mock==3.6.1
PyYAML==5.4.1
